requirements.txt
python-dotenv==1.0.1
pypdf==4.3.0
langchain_community==0.2.7
faiss-cpu==1.8.0.post1
streamlit==1.36.0
langchain==0.2.8
sentence-transformers==3.0.1
langchain_huggingface==0.0.3